// generated/src/app/api-models/aucls-job-determinant.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuclsJobDeterminant {
  seqAuclsJobId: string;
  authsColumnName: string;
  operator: string;
  fromValue: string;
  thruValue: string;
  authsColumnTypeCode: number;
  columnOccurrence: number;
  columnType: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}